+++ common/src/main/java/dev/ftb/mods/ftbessentials/command/TeleportCommands.java
@@ -22,7 +22,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.tags.BiomeTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.Mth;
 import net.minecraft.world.level.Level;
@@ -54,6 +53,13 @@ public class TeleportCommands {
 			dispatcher.register(Commands.literal("spawn")
 					.requires(FTBEConfig.SPAWN)
 					.executes(context -> spawn(context.getSource().getPlayerOrException()))
+			);
+		}
+
+		if (FTBEConfig.PLAYER_SPAWN.isEnabled()) {
+			dispatcher.register(Commands.literal("playerspawn")
+					.requires(FTBEConfig.PLAYER_SPAWN)
+					.executes(context -> playerSpawn(context.getSource().getPlayerOrException()))
 			);
 		}
 
@@ -94,6 +100,11 @@ public class TeleportCommands {
 		try {
 			ServerPlayer player = source.getPlayerOrException();
 
+			TeleportPos.TeleportResult blacklistResult = new TeleportPos(player).checkDimensionBlacklist(player);
+			if (!blacklistResult.isSuccess()) {
+				return blacklistResult.runCommand(player);
+			}
+
 			BlockHitResult res = BlockUtil.getFocusedBlock(player, player.getServer().getPlayerList().getViewDistance() * 16)
 					.orElseThrow(() -> new IllegalArgumentException("Not looking at a block"));
 			// want to land the player on top of the focused block, so scan up as far as needed
@@ -132,7 +143,7 @@ public class TeleportCommands {
 		}).orElse(0);
 	}
 
-	public static int spawn(ServerPlayer player) {
+	public static int playerSpawn(ServerPlayer player) {
 		return FTBEPlayerData.getOrCreate(player).map(data -> {
 			ServerLevel level = player.server.getLevel(player.getRespawnDimension());
 			if (level == null) {
@@ -143,10 +154,23 @@ public class TeleportCommands {
 		}).orElse(0);
 	}
 
+	public static int spawn(ServerPlayer player) {
+		return FTBEPlayerData.getOrCreate(player).map(data -> {
+			ServerLevel level = player.server.getLevel(Level.OVERWORLD);
+			return level == null ? 0 : data.spawnTeleporter.teleport(player, p -> new TeleportPos(level, level.getSharedSpawnPos(), level.getSharedSpawnAngle(), 0F)).runCommand(player);
+		}).orElse(0);
+	}
+
 	public static int rtp(ServerPlayer player) {
-		if (!player.hasPermissions(2) && !DimensionFilter.isDimensionOK(player.level().dimension())) {
-			player.displayClientMessage(Component.literal("You may not use /rtp in this dimension!").withStyle(ChatFormatting.RED), false);
-			return 0;
+		if (!player.hasPermissions(2)) {
+			TeleportPos.TeleportResult blacklistResult = new TeleportPos(player).checkDimensionBlacklist(player);
+			if (!blacklistResult.isSuccess()) {
+				return blacklistResult.runCommand(player);
+			}
+			if (!DimensionFilter.isRtpDimensionOK(player.level().dimension())) {
+				player.displayClientMessage(Component.literal("You may not use /rtp in this dimension!").withStyle(ChatFormatting.RED), false);
+				return 0;
+			}
 		}
 		return FTBEPlayerData.getOrCreate(player).map(data -> data.rtpTeleporter.teleport(player, p -> {
 					p.displayClientMessage(Component.literal("Looking for random location..."), false);
@@ -207,19 +231,29 @@ public class TeleportCommands {
 	public static int tpLast(ServerPlayer player, GameProfile to) {
 		ServerPlayer toPlayer = player.server.getPlayerList().getPlayer(to.getId());
 		if (toPlayer != null) {
-			FTBEPlayerData.addTeleportHistory(player);
-			new TeleportPos(toPlayer).teleport(player);
-			return 1;
+			TeleportPos.TeleportResult result = new TeleportPos(toPlayer).teleport(player);
+			if (result.isSuccess()) {
+				FTBEPlayerData.addTeleportHistory(player);
+				return 1;
+			}
+			return result.runCommand(player);
 		}
 		// dest player not online; teleport to where they were last seen
 		return FTBEPlayerData.getOrCreate(to).map(dataTo -> {
-			FTBEPlayerData.addTeleportHistory(player);
-			dataTo.getLastSeenPos().teleport(player);
-			return 1;
+			TeleportPos.TeleportResult result = dataTo.getLastSeenPos().teleport(player);
+			if (result.isSuccess()) {
+				FTBEPlayerData.addTeleportHistory(player);
+				return 1;
+			}
+			return result.runCommand(player);
 		}).orElse(0);
 	}
 
 	public static int tpx(ServerPlayer player, ServerLevel to) {
+		TeleportPos.TeleportResult blacklistResult = new TeleportPos(to, player.blockPosition(), player.getYRot(), player.getXRot()).checkDimensionBlacklist(player);
+		if (!blacklistResult.isSuccess()) {
+			return blacklistResult.runCommand(player);
+		}
 		player.teleportTo(to, player.getX(), player.getY(), player.getZ(), player.getYRot(), player.getXRot());
 		return 1;
 	}